requirements.txt
appdirs==1.4.4
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
cssselect==1.1.0
et-xmlfile==1.1.0
fake-useragent==0.1.11
h11==0.13.0
idna==3.3
importlib-metadata==4.11.4
lxml==4.8.0
numpy==1.22.4
openpyxl==3.0.10
outcome==1.1.0
parse==1.19.0
pycparser==2.21
pydantic==1.9.1
pyee==8.2.2
pyOpenSSL==22.0.0
pyppeteer==1.0.2
pyquery==1.4.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
requests-html==0.10.0
selenium==4.2.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tqdm==4.64.0
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.2.0
urllib3==1.26.9
w3lib==1.22.0
webdriver-manager==3.7.0
websockets==10.3
WooCommerce==3.0.0
wsproto==1.1.0
zipp==3.8.0
